% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_decode}
\alias{base64_decode}
\alias{base64_encode}
\title{Base64 Encoding and Decoding}
\usage{
base64_decode(x)

base64_encode(x)
}
\arguments{
\item{x}{Raw vector to encode / decode.}
}
\value{
Raw vector, result of the encoding / decoding.
}
\description{
Base64 Encoding and Decoding
}
