% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{curl_fds}
\alias{curl_fds}
\title{Create a pollable object from a curl multi handle's file descriptors}
\usage{
curl_fds(fds)
}
\arguments{
\item{fds}{A list of file descriptors, as returned by
\code{\link[curl:multi_fdset]{curl::multi_fdset()}}.}
}
\value{
Pollable object, that be used with \code{\link[=poll]{poll()}} directly.
}
\description{
Create a pollable object from a curl multi handle's file descriptors
}
