% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{exec}
\alias{exec}
\alias{exec-package}
\title{Execute and Control System Processes}
\description{
Portable tools to run system processes, get their process identifier,
cancel and restart them.
}

