% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_pid_from_file}
\alias{get_pid_from_file}
\alias{get_pid_from_file_windows}
\alias{get_pid_from_file_unix}
\title{Parse a PID from an output file written by \code{get_my_pid()}}
\usage{
get_pid_from_file(pidfile, cmdfile)

get_pid_from_file_windows(pidfile, cmdfile)

get_pid_from_file_unix(pidfile, cmdfile)
}
\arguments{
\item{pidfile}{Name of the pid file, same as the argument to
\code{get_my_pid_file()}.}

\item{cmdfile}{Name of the script file, we use this as an ID on
windows, to find the process among the children of the R process,
because windows shells cannot report their own process id.}
}
\description{
The format of the file is platform dependent.
}
\seealso{
Other pid queries: \code{\link{get_my_pid_code}}
}
\keyword{internal}
