% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handles.R
\name{handle_create}
\alias{handle_create}
\alias{handle_describe}
\title{OS handles}
\usage{
handle_create(desc)

handle_describe(handle)
}
\arguments{
\item{desc}{An integer or string, the description of an OS handle.}

\item{handle}{A \code{processx_handle} S3 object, an external pointer to
an OS handle.}
}
\value{
\code{handle_create()} returns a \code{processx_handle} S3 object, which
is an external pointer to an OS handle.

\code{handle_describe()} returns a string.
}
\description{
Helper functions to create R objects that refer to OS specific
handles. These are \code{HANDLE}s on Windows and file descriptors on Unix-like
systems.
}
\details{
\code{handle_create()} creates a handle object, from a description.
This is an external pointer, pointing to a \code{HANDLE} (Windows) or an
\code{int} (file descriptor, Unix).

\code{handle_describe()} creates an R string that is a description of the
OS handle.

These functions are for advanced users, and are considered as
experimental currently. They might change in future processx versions.
}
