% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll}
\alias{poll}
\title{Poll for process I/O or termination}
\usage{
poll(processes, ms)
}
\arguments{
\item{processes}{A list of \code{process} objects to wait on. If this is a
named list, then the returned list will have the same names. This
simplifies the identification of the processes. If an empty list,
then the}

\item{ms}{Integer scalar, a timeout for the polling, in milliseconds.
Supply -1 for an infitite timeout, and 0 for not waiting at all.}
}
\value{
A list of character vectors of length two. There is one list
element for each process, in the same order as in the input list.
The character vectors' elements are named \code{output} and \code{error} and
their possible values are: \code{nopipe}, \code{ready}, \code{timeout}, \code{closed},
\code{silent}. See details about these below.
}
\description{
Wait until one of the specified processes produce standard output
or error, terminates, or a timeout occurs.
}
\section{Explanation of the return values}{

\itemize{
\item \code{nopipe} means that the stdout or stderr from this process was not
captured.
\item \code{ready} means that stdout or stderr from this process are ready to
read from. Note that end-of-file on these outputs also triggers
\code{ready}.
\item timeout`: the processes are not ready to read from and a timeout
happened.
\item \code{closed}: the connection was already closed, before the polling
started.
\item \code{silent}: the connection is not ready to read from, but another
connection was.
}
}

\section{Known issues}{


You cannot wait on the termination of a process directly. It is only
signalled through the closed stdout and stderr pipes. This means that
if both stdout and stderr are ignored or closed for a process, then you
will not be notified when it exits.
}

\examples{
# TODO
}
