% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processx-package.R
\docType{package}
\name{processx-package}
\alias{processx}
\alias{processx-package}
\title{processx: Execute and Control System Processes}
\description{
Tools to run system processes in the background. It can check if a background process is running; wait on a background process to finish; get the exit status of finished processes; kill background processes. It can read the standard output and error of the processes, using non-blocking connections. 'processx' can poll a process for standard output or error, with a timeout. It can also poll several processes at once.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://processx.r-lib.org}
  \item \url{https://github.com/r-lib/processx}
  \item Report bugs at \url{https://github.com/r-lib/processx/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Winston Chang
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Mango Solutions [copyright holder, funder]
}

}
\keyword{internal}
