% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{conn_create_fd}
\alias{conn_create_fd}
\alias{conn_create_pipepair}
\alias{conn_read_chars}
\alias{read_chars.processx_connection}
\alias{conn_read_lines}
\alias{conn_read_lines.processx_connection}
\alias{conn_is_incomplete}
\alias{conn_is_incomplete.processx_connection}
\alias{conn_write}
\alias{conn_write.processx_connection}
\title{Processx connections}
\usage{
conn_create_fd(fd, encoding = "")

conn_create_pipepair(encoding = "")

conn_read_chars(con, n = -1)

read_chars.processx_connection(con, n = -1)

conn_read_lines(con, n = -1)

conn_read_lines.processx_connection(con, n = -1)

conn_is_incomplete(con)

conn_is_incomplete.processx_connection(con)

conn_write(con, str, sep = "\\n", encoding = "")

conn_write.processx_connection(con, str, sep = "\\n", encoding = "")
}
\arguments{
\item{fd}{Integer scalar, a Unix file descriptor.}

\item{encoding}{Encoding of the readable connection when reading.
Encoding to re-encode \code{str} into when writing.}

\item{con}{Processx connection object.}

\item{n}{Number of characters or lines to read. -1 means all available
characters or lines.}

\item{str}{Character or raw vector to write.}

\item{sep}{Separator to use if \code{str} is a character vector. Ignored if
\code{str} is a raw vector.}
}
\description{
These functions are currently experimental and will change
in the future. Note that processx connections are  \emph{not}
compatible with R's built-in connection system.

\code{conn_create_pipepair()} creates a pair of connected connections, the
first one is writeable, the second one is readable.

\code{conn_read_chars()} reads UTF-8 characters from the connections. If the
connection itself is not UTF-8 encoded, it re-encodes it.

\code{conn_read_lines()} reads lines from a connection.

\code{conn_is_incomplete()} returns \code{FALSE} if the connection surely has no
more data.

\code{conn_write()} writes a character or raw vector to the connection.
It might not be able to write all bytes into the connection, in which
case it returns the leftover bytes in a raw vector. Call \code{conn_write()}
again with this raw vector.
}
\details{
\code{conn_create_fd()} creates a connection from a file descriptor.
}
