% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{serialize_to_raw}
\alias{serialize_to_raw}
\title{Serialize an R object into a raw vector}
\usage{
serialize_to_raw(x)
}
\arguments{
\item{x}{Object to serialize.}
}
\value{
Raw vector, serialized \code{x}.
}
\description{
Compared to \code{\link[=serialize]{serialize()}}, it uses \code{xdr = FALSE}, \code{ascii = FALSE}, and
\code{version} is 3 for R >= 3.5.0 and 2 otherwise.
}
\examples{
x <- 1:10
serialize_to_raw(x)
unserialize_from_raw(serialize_to_raw(x))
}
\seealso{
Other serialization: \code{\link{unserialize_from_raw}}
}
\concept{serialization}
