% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{unserialize_from_raw}
\alias{unserialize_from_raw}
\title{Unserialize an R object from a raw vector}
\usage{
unserialize_from_raw(x)
}
\arguments{
\item{x}{Raw vector.}
}
\value{
Unserialized object.
}
\description{
Unserialize an R object from a raw vector
}
\examples{
x <- 1:10
serialize_to_raw(x)
unserialize_from_raw(serialize_to_raw(x))
}
\seealso{
Other serialization: \code{\link{serialize_to_raw}}
}
\concept{serialization}
