% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{default_pty_options}
\alias{default_pty_options}
\title{Default options for pseudo terminals (ptys)}
\usage{
default_pty_options()
}
\value{
Named list of default values of pty options.

Options and default values:
\itemize{
\item \code{echo} whether to keep the echo on the terminal. \code{FALSE} turns echo
off.
\item \code{rows} the (initial) terminal size, number of rows.
\item \code{cols} the (initial) terminal size, number of columns.
}
}
\description{
Default options for pseudo terminals (ptys)
}
