% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{process_initialize}
\alias{process_initialize}
\title{Start a process}
\usage{
process_initialize(self, private, command, args, commandline, stdout, stderr,
  cleanup, echo_cmd, windows_verbatim_args, windows_hide_window)
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{command}{Command to run, string scalar.}

\item{args}{Command arguments, character vector.}

\item{commandline}{Alternative to command + args.}

\item{stdout}{Standard output, FALSE to ignore, TRUE for temp file.}

\item{stderr}{Standard error, FALSE to ignore, TRUE for temp file.}

\item{cleanup}{Kill on GC?}

\item{echo_cmd}{Echo command before starting it?}
}
\description{
Start a process
}
\keyword{internal}
